refactor: 덱 클래스 싱글턴으로 변경
