+++ src/main/kotlin/blackjack/Deck.kt
@@ -1,23 +1,21 @@
 package blackjack
 
-class Deck {
-    companion object {
-        const val INITIAL_HAND_OUT_CARD_COUNT = 2
-        private val CARDS = generateCards()
+object Deck {
+    const val INITIAL_HAND_OUT_CARD_COUNT = 2
+    private val CARDS = generateCards()
 
-        fun draw(): Card {
-            return CARDS.removeFirstOrNull() ?: throw IllegalStateException("카드가 더 없습니다.")
-        }
+    fun draw(): Card {
+        return CARDS.removeFirstOrNull() ?: throw IllegalStateException("카드가 더 없습니다.")
+    }
 
-        private fun generateCards(): MutableList<Card> {
-            val cards = mutableListOf<Card>()
+    private fun generateCards(): MutableList<Card> {
+        val cards = mutableListOf<Card>()
 
-            Shape.entries.forEach { shape ->
-                Number.entries.forEach { number ->
-                    cards.add(Card(shape, number))
-                }
+        Shape.entries.forEach { shape ->
+            Number.entries.forEach { number ->
+                cards.add(Card(shape, number))
             }
-            return cards.shuffled().toMutableList()
         }
+        return cards.shuffled().toMutableList()
     }
 }
