+++ src/main/kotlin/blackjack/controller/BlackjackController.kt
@@ -2,7 +2,7 @@ package blackjack.controller
 
 import blackjack.Dealer
 import blackjack.Deck
-import blackjack.Deck.Companion.INITIAL_HAND_OUT_CARD_COUNT
+import blackjack.Deck.INITIAL_HAND_OUT_CARD_COUNT
 import blackjack.GameManager
 import blackjack.Player
 import blackjack.view.InputView
